+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Add `Duration` capable of parsing human-readable duration formats ([#647])
+
+[#647]: https://github.com/stackabletech/operator-rs/pull/647
+
 ## [0.48.0] - 2023-08-18
 
 ### Added