+++ CHANGELOG.md
@@ -16,8 +16,10 @@ All notable changes to this project will be documented in this file.
 
 - Add `PodBuilder::termination_grace_period_seconds` ([#641]).
 - Add support for adding `lifecycle`s to `ContainerBuilder` ([#641]).
+- Add support for tls pkcs12 password to secret operator volume builder ([#645]).
 
 [#641]: https://github.com/stackabletech/operator-rs/pull/641
+[#645]: https://github.com/stackabletech/operator-rs/pull/645
 
 ## [0.47.0] - 2023-08-16